# Using input.place() in GetExpectedKernel in slice_op